requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
MouseInfo==0.1.3
openai==1.97.1
pillow==11.3.0
pytweening==1.2.0
pywin32==311
screen_brightness_control==0.24.2
sniffio==1.3.1
tk==0.1.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
WMI==1.5.1
pyautogui
google-generativeai